Revert "Merge branch 'Feat/new-contract' into dev"
This reverts commit 47158ba4d4223bd450a2c4b29d8829a96d8545fc, reversing
changes made to 544f4de4d6871ef5ca9fa7f462c10086a88c3366.

+++ src/main/java/com/alia/back_end_service/jpa/league/LeagueEntity.java
@@ -16,7 +16,6 @@ import java.util.List;
 @Table(name = "leagues")
 @Getter
 @Setter
-
 public class LeagueEntity {
 
     @Id
@@ -33,7 +32,7 @@ public class LeagueEntity {
     private OffsetDateTime end_time;
 
     @Column(name = "time_match", nullable = false)
-    private Long time_match;
+    private Integer time_match;
 
     @Column(name = "number_match", nullable = false)
     private Integer number_match;
@@ -47,7 +46,6 @@ public class LeagueEntity {
             joinColumns = @JoinColumn(name = "league_id"),
             inverseJoinColumns = @JoinColumn(name = "bot_id")
     )
-
     private List<BotEntity> bots = new ArrayList<>();
 
     @OneToMany(mappedBy = "league", cascade = CascadeType.ALL, orphanRemoval = true, fetch = FetchType.LAZY)
